Ingestor/requirements.txt
# Core dependencies for multi-format ingestion
python-docx>=1.1.0
pypdf>=4.2.0
pdfplumber>=0.11.0
pandas>=2.2.0
openpyxl>=3.1.0
xlrd>=2.0.1

# Legacy .doc extraction on Windows (uses Microsoft Word COM automation)
pywin32>=306; platform_system == "Windows"

# Legacy .doc fallback extractor for non-Windows environments
textract>=1.6.5; platform_system != "Windows"
